use * as a nameservice wildcard
Having to specify the name service is annoying if the machine where
rock-webapp runs changes w.r.t. where the components run and/or if
the IPs change.

For the simple (but actually pretty common) case where a rock-webapp
process serves only one rock system, use '*' to do relative resolution.